up: - replaced float with decimal as this type takes effect in db - added cspell file to sync dictionary across the project - added address geolocation as preferable field

+++ app/Models/Company.php
@@ -21,7 +21,7 @@ class Company extends Model
     ];
 
     protected $casts = [
-        'tax_value' => 'float',
+        'tax_value' => 'decimal',
     ];
 
     public function users(): BelongsToMany
